Исправил доп баги.

+++ src/main/java/org/example/project/service/comparator/impl/RootCropWeightComparator.java
@@ -1,15 +1,12 @@
 package org.example.project.service.comparator.impl;
 
-import org.example.project.model.impl.Book;
-import org.example.project.model.impl.Car;
 import org.example.project.model.impl.RootCrop;
-import org.example.project.service.comparator.ProductComparable;
 
 import java.util.Comparator;
 
 public class RootCropWeightComparator implements Comparator<RootCrop> {
     @Override
     public int compare(RootCrop rootCrop1, RootCrop rootCrop2) {
-        return Integer.compare(rootCrop1.getWeight(), rootCrop2.getWeight());
+        return Double.compare(rootCrop1.getWeight(), rootCrop2.getWeight());
     }
 }
